feat: Add PraxisTeam component and update Home page
The commit adds the PraxisTeam component to the Home page of the vet-clinic project. This component is responsible for rendering the team members on the website. It imports the necessary dependencies and displays the team members using the TeamCard component. This change enhances the website by showcasing the team members and their information.